# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис для проверки доступности сайтов",
        "title": "watcher",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/admin/stat": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Return most fasts site",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Administration"
                ],
                "summary": "Return statistic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetMinResponse"
                        }
                    }
                }
            }
        },
        "/stat/max": {
            "get": {
                "description": "Return most slow site",
                "produces": [
                    "application/json"
                ],
                "summary": "Return most slow site",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetMaxResponse"
                        }
                    }
                }
            }
        },
        "/stat/min": {
            "get": {
                "description": "Return most fasts site",
                "produces": [
                    "application/json"
                ],
                "summary": "Return most fasts site",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetMinResponse"
                        }
                    }
                }
            }
        },
        "/stat/{id}/site": {
            "get": {
                "description": "Return most fasts site",
                "produces": [
                    "application/json"
                ],
                "summary": "Return most fasts site",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Site ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.GetSiteStatResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.GetMaxResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "seconds": {
                    "type": "number"
                }
            }
        },
        "handler.GetMinResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "seconds": {
                    "type": "number"
                }
            }
        },
        "handler.GetSiteStatResponse": {
            "type": "object",
            "properties": {
                "IsAlive": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "seconds": {
                    "type": "number"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}